refactor(core): P1-7 提取为独立 retrieval_evaluation 模块，P1 不依赖 #181 文件
- 恢复 cerebellum_optimizer.py / events.py 为 #181 版本（P1 提交不修改 #181 新文件）
- 新增 core/loop/retrieval_evaluation.py: split_held_out_qa / evaluate_retrieval / cerebellum 适配器 (自包含 _import_cerebellum)
- test_retrieval_evaluation.py 改引 retrieval_evaluation
