+++ tests/test_retrieval_evaluation.py
@@ -2,8 +2,8 @@
 
 Lesson 15's weak-evaluation traps: (a) eval sets built from the very documents
 being indexed inflate scores (same-source), and (b) exact-string scoring has
-zero tolerance for paraphrase. These tests pin the two fixes — held-out QA
-sources and semantic (embedding-cosine) hit scoring — on top of cerebellum's
+zero tolerance for paraphrase. These tests pin the two fixes 鈥?held-out QA
+sources and semantic (embedding-cosine) hit scoring 鈥?on top of cerebellum's
 existing MRR benchmark.
 """
 
@@ -16,7 +16,7 @@ ROOT = Path(__file__).resolve().parents[1]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from core.loop import cerebellum_optimizer as co
+from core.loop import retrieval_evaluation as re
 
 # ---- held-out QA split ------------------------------------------------------
 
@@ -27,7 +27,7 @@ def test_split_held_out_isolates_sources():
         {"content": "alpha fact two", "source": "alpha"},
         {"content": "beta fact one", "source": "beta"},
     ]
-    qa, indexed = co.split_held_out_qa(entries, {"alpha"})
+    qa, indexed = re.split_held_out_qa(entries, {"alpha"})
     assert len(qa) == 2
     assert all(q["source"] == "alpha" for q in qa)
     assert indexed == {"beta"}
@@ -35,7 +35,7 @@ def test_split_held_out_isolates_sources():
 
 def test_split_keeps_query_prefix_and_gold():
     entries = [{"content": "A very long durable fact worth remembering", "source": "x"}]
-    qa, indexed = co.split_held_out_qa(entries, {"x"})
+    qa, indexed = re.split_held_out_qa(entries, {"x"})
     assert indexed == set()
     assert len(qa) == 1
     assert qa[0]["gold"] == "A very long durable fact worth remembering"
@@ -50,14 +50,14 @@ def test_split_skips_blank_and_malformed():
         "not-a-dict",
         {"content": "valid", "source": "y"},
     ]
-    qa, _indexed = co.split_held_out_qa(entries, {"x", "y"})
+    qa, _indexed = re.split_held_out_qa(entries, {"x", "y"})
     assert len(qa) == 1
     assert qa[0]["gold"] == "valid"
 
 
 def test_no_hold_out_keeps_all_indexed():
     entries = [{"content": "fact", "source": "a"}]
-    qa, indexed = co.split_held_out_qa(entries, set())
+    qa, indexed = re.split_held_out_qa(entries, set())
     assert qa == []
     assert indexed == {"a"}
 
@@ -66,7 +66,7 @@ def test_no_hold_out_keeps_all_indexed():
 
 
 def _fake_embed(text: str) -> list[float] | None:
-    """Deterministic toy embedder: bag-of-tokens → vector, so cosine reflects
+    """Deterministic toy embedder: bag-of-tokens 鈫?vector, so cosine reflects
     token overlap (a cheap stand-in for paraphrase tolerance)."""
 
     tokens = {w for w in str(text).lower().split() if w.isalnum()}
@@ -82,10 +82,10 @@ def _search_returning(contents: list[str]):
 
 
 def test_semantic_hit_counts_paraphrase():
-    # Gold and hit differ in wording but share tokens → cosine ≥ threshold.
+    # Gold and hit differ in wording but share tokens 鈫?cosine 鈮?threshold.
     qa = [{"query": "tell me about alpha facts", "gold": "alpha fact details"}]
     search = _search_returning(["the alpha facts explained", "unrelated doc"])
-    metrics = co.evaluate_retrieval(
+    metrics = re.evaluate_retrieval(
         qa, search_fn=search, embed_fn=_fake_embed, top_k=5
     )
     assert metrics["recall@1"] == 1.0
@@ -97,7 +97,7 @@ def test_semantic_hit_counts_paraphrase():
 def test_semantic_hit_at_second_position_ranked():
     qa = [{"query": "tell me about alpha facts", "gold": "alpha fact details"}]
     search = _search_returning(["unrelated doc", "the alpha facts explained"])
-    metrics = co.evaluate_retrieval(
+    metrics = re.evaluate_retrieval(
         qa, search_fn=search, embed_fn=_fake_embed, top_k=5
     )
     assert metrics["recall@1"] == 0.0
@@ -108,7 +108,7 @@ def test_semantic_hit_at_second_position_ranked():
 def test_semantic_scoring_rejects_unrelated():
     qa = [{"query": "alpha question", "gold": "alpha gold answer"}]
     search = _search_returning(["completely unrelated text"])
-    metrics = co.evaluate_retrieval(qa, search_fn=search, embed_fn=_fake_embed)
+    metrics = re.evaluate_retrieval(qa, search_fn=search, embed_fn=_fake_embed)
     assert metrics["recall@1"] == 0.0
     assert metrics["mrr"] == 0.0
     assert metrics["per_query"][0]["rank"] == 0
@@ -117,17 +117,17 @@ def test_semantic_scoring_rejects_unrelated():
 def test_weak_mode_without_embedder_is_explicit():
     qa = [{"query": "q", "gold": "exact phrase"}]
     search = _search_returning(["exact phrase", "other"])
-    metrics = co.evaluate_retrieval(qa, search_fn=search, embed_fn=None)
+    metrics = re.evaluate_retrieval(qa, search_fn=search, embed_fn=None)
     assert metrics["weak"] is True  # explicitly flagged, not silent
     assert metrics["recall@1"] == 1.0  # exact-substring still counts
-    # Exact match at position 2 → rank 2.
+    # Exact match at position 2 鈫?rank 2.
     search2 = _search_returning(["other", "exact phrase"])
-    metrics2 = co.evaluate_retrieval(qa, search_fn=search2, embed_fn=None)
+    metrics2 = re.evaluate_retrieval(qa, search_fn=search2, embed_fn=None)
     assert metrics2["per_query"][0]["rank"] == 2
 
 
 def test_empty_qa_set_returns_zeros():
-    metrics = co.evaluate_retrieval([], search_fn=_search_returning([]))
+    metrics = re.evaluate_retrieval([], search_fn=_search_returning([]))
     assert metrics["queries"] == 0
     assert metrics["recall@1"] == 0.0 and metrics["mrr"] == 0.0
 
@@ -137,7 +137,7 @@ def test_search_failure_counts_as_miss():
         raise RuntimeError("store down")
 
     qa = [{"query": "q", "gold": "g"}]
-    metrics = co.evaluate_retrieval(qa, search_fn=_boom, embed_fn=_fake_embed)
+    metrics = re.evaluate_retrieval(qa, search_fn=_boom, embed_fn=_fake_embed)
     assert metrics["recall@1"] == 0.0
     assert metrics["per_query"][0]["rank"] == 0
 
@@ -146,11 +146,11 @@ def test_search_failure_counts_as_miss():
 
 
 def test_search_adapter_missing_cerebellum_returns_empty(monkeypatch, tmp_path):
-    monkeypatch.setattr(co, "_CEREBELLUM_EVOLUTION", tmp_path / "nope.py")
-    search = co.cerebellum_search_adapter()
+    monkeypatch.setattr(re, "_CEREBELLUM_EVOLUTION", tmp_path / "nope.py")
+    search = re.cerebellum_search_adapter()
     assert search("anything", 5) == []
 
 
 def test_embed_adapter_missing_cerebellum_returns_none(monkeypatch, tmp_path):
-    monkeypatch.setattr(co, "_CEREBELLUM_EVOLUTION", tmp_path / "nope.py")
-    assert co.cerebellum_embed_adapter() is None
+    monkeypatch.setattr(re, "_CEREBELLUM_EVOLUTION", tmp_path / "nope.py")
+    assert re.cerebellum_embed_adapter() is None
